fix: address PRRT_kwDOSJAM6s6KzZLI oversized shard
