+++ plans/PRRT_kwDOSJAM6s6KzZLI_VALIDATION.md
@@ -1,0 +1,37 @@
+# PRRT_kwDOSJAM6s6KzZLI Validation
+
+Plan reference: `PRRT_kwDOSJAM6s6KzZLI_PLAN.md`
+
+## Requirement Status
+
+- Keep the change scoped to splitting the oversized test shard: Complete.
+  Only part 009, the new adjacent part 011, and the required plan/validation
+  notes were changed.
+- Preserve existing tests and assertions: Complete.
+  The final direct `quality_methods` helper tests were moved intact into part
+  011 with local imports/constants/helpers.
+- Keep every touched file under the line limit: Complete.
+  Part 009 is now 1,414 lines and part 011 is 166 lines.
+- Run focused checks only: Complete.
+  Broad AWF/GitHub validation is intentionally left to AWF after agent
+  completion.
+
+## Evidence
+
+- Initial focused guard before implementation:
+  `uv run --python 3.12 --extra dev pytest tests/unit/test_core_decomposition_maintainability.py::test_first_party_code_files_stay_under_line_limit -q`
+  failed with part 009 reported at 1,564 lines.
+- After implementation:
+  `uv run --python 3.12 --extra dev pytest tests/unit/test_core_decomposition_maintainability.py::test_first_party_code_files_stay_under_line_limit -q`
+  passed (`1 passed`).
+- Touched shards:
+  `uv run --python 3.12 --extra dev pytest tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_009.py tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_011.py -q`
+  passed (`28 passed`).
+- Focused lint:
+  `uv run --python 3.12 --extra dev ruff check tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_009.py tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_011.py`
+  passed.
+
+## Gaps
+
+None. Full validation, coverage, and PR gating are managed by AWF/GitHub after
+this agent phase.
